pre-final release

+++ app.py
@@ -1,6 +1,7 @@
 import os
-import random
+import time
 import requests
+import json
 from lxml import etree
 from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
@@ -10,6 +11,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from webdriver_manager.chrome import ChromeDriverManager
 from slugify import slugify
+from datetime import datetime
 
 class BandcampDownload:
 
@@ -47,26 +49,37 @@ class BandcampDownload:
     def _get_bc_show_dl_link(self, show_link):
         _chrome_browser_handle = self._get_chrome_browser_handle()
         _chrome_browser_handle.get(show_link)
-        wait = WebDriverWait(_chrome_browser_handle, timeout=30)
+        wait = WebDriverWait(_chrome_browser_handle, timeout=60)
         download_ready = wait.until(EC.presence_of_element_located((By.XPATH, "//*[@class='item-button' and @href]")))
         download_url = download_ready.get_attribute("href")
-        show_name = _chrome_browser_handle.find_element(By.CLASS_NAME, "title").text
-        return (download_url, show_name)
+        blob = _chrome_browser_handle.find_element(By.ID, "pagedata").get_attribute("data-blob")
+        parse_blob = json.loads(blob)
+        show_title = parse_blob["digital_items"][0]["title"]
+        show_slug = parse_blob["digital_items"][0]["url_hints"]["slug"]
+        return (download_url, show_title, show_slug)
 
     def download_show(self, show_link):
-        download_url, show_name = self._get_bc_show_dl_link(show_link=show_link)
-        print(f'Show Name: {show_name}')
-        show_name_slug = slugify(show_name)
-        print(f'Show Name SLUG: {show_name_slug}')
+        show = self._get_bc_show_dl_link(show_link=show_link)
+        download_url = show[0]
+        show_title = show[1]
+        show_slug = show[2]
+        print(f'Show Title: {show_title}')
+        print(f'Show Name SLUG: {show_slug}')
+        file_path = os.path.join("F:\\", f'{show_slug}.zip')
         print(f'Download URL: {download_url}')
-        download_request = requests.get(download_url, allow_redirects=True)
-        file_path = os.path.join("D:\MUSIC\FLAC\STS9", f'{show_name_slug}.zip')
-        if open(file_path, 'wb').write(download_request.content):
-            print(f'Downloaded show to path: {file_path}')
+        if not os.path.isfile(file_path):
+            download_request = requests.get(download_url, allow_redirects=True)
+            if open(file_path, 'wb').write(download_request.content):
+                print(f'Downloaded show to path: {file_path}')
+                return True
+        print(f'Already downloaded - skipping...')
+        return False
 
 
 if __name__ == "__main__":
 
     bc_dl = BandcampDownload(bc_html="bandcamp_dump.html")
     print(f'Found {len(bc_dl.dl_links)} shows for download')
-    bc_dl.download_show(show_link=bc_dl.dl_links[random.randint(0, len(bc_dl.dl_links))])
+    for show_link in bc_dl.dl_links:
+        data = bc_dl.download_show(show_link=show_link)
+        time.sleep(30)
